Refactor Apollo stuffs to separate projects
