+++ modules/apollo-cache-inmemory/src/main/scala/anduin/scalajs/apollo/cache/ApolloInMemoryCacheOptions.scala
@@ -1,0 +1,7 @@
+package anduin.scalajs.apollo.cache
+
+import scala.scalajs.js
+
+final case class ApolloInMemoryCacheOptions(
+  addTypename: js.UndefOr[Boolean] = js.undefined
+)
