+++ modules/apollo-cache-inmemory/src/main/scala/anduin/scalajs/apollo/cache/internal/ApolloInMemoryCache.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.cache.internal
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+@JSImport("apollo-cache-inmemory", "InMemoryCache", "ApolloCacheInmemory.InMemoryCache")
+@js.native
+class ApolloInMemoryCache(options: ApolloInMemoryCacheOptions) extends ApolloCache
