+++ modules/apollo-cache-inmemory/src/main/scala/anduin/scalajs/apollo/cache/internal/ApolloInMemoryCacheOptions.scala
@@ -1,0 +1,7 @@
+package anduin.scalajs.apollo.cache.internal
+
+import scala.scalajs.js
+
+final class ApolloInMemoryCacheOptions(
+  val addTypename: js.UndefOr[Boolean]
+) extends js.Object
