+++ modules/apollo-cache/src/main/scala/anduin/scalajs/apollo/cache/ApolloCache.scala
@@ -1,0 +1,6 @@
+package anduin.scalajs.apollo.cache
+
+trait ApolloCache {
+  type Raw <: internal.ApolloCache
+  val raw: Raw
+}
