+++ modules/apollo-cache/src/main/scala/anduin/scalajs/apollo/cache/internal/ApolloCache.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.cache.internal
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+@JSImport("apollo-cache", "ApolloCache", "ApolloCache.ApolloCache")
+@js.native
+class ApolloCache extends js.Object
