+++ modules/apollo-client/src/main/scala/anduin/scalajs/apollo/client/ApolloClientOptions.scala
@@ -1,0 +1,9 @@
+package anduin.scalajs.apollo.client
+
+import anduin.scalajs.apollo.cache.ApolloCache
+import anduin.scalajs.apollo.link.ApolloLink
+
+final case class ApolloClientOptions(
+  link: ApolloLink,
+  cache: ApolloCache
+)
