+++ modules/apollo-client/src/main/scala/anduin/scalajs/apollo/client/internal/ApolloClient.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.client.internal
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+@JSImport("apollo-client", "ApolloClient", "ApolloClient.ApolloClient")
+@js.native
+class ApolloClient(options: ApolloClientOptions) extends js.Object
