+++ modules/apollo-client/src/main/scala/anduin/scalajs/apollo/client/internal/ApolloClientOptions.scala
@@ -1,0 +1,11 @@
+package anduin.scalajs.apollo.client.internal
+
+import anduin.scalajs.apollo.cache.internal.ApolloCache
+import anduin.scalajs.apollo.link.internal.ApolloLink
+
+import scala.scalajs.js
+
+final class ApolloClientOptions(
+  val link: ApolloLink,
+  val cache: ApolloCache
+) extends js.Object
