+++ modules/apollo-link-http/src/main/scala/anduin/scalajs/apollo/link/ApolloHttpLinkOptions.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.link
+
+import anduin.scalajs.fetch.Fetch
+
+final case class ApolloHttpLinkOptions(
+  uri: String,
+  fetch: Fetch
+)
