+++ modules/apollo-link-http/src/main/scala/anduin/scalajs/apollo/link/internal/ApolloHttpLink.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.link.internal
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+@JSImport("apollo-link-http", "HttpLink", "ApolloLinkHttp.HttpLink")
+@js.native
+class ApolloHttpLink(options: ApolloHttpLinkOptions) extends ApolloLink
