+++ modules/apollo-link-http/src/main/scala/anduin/scalajs/apollo/link/internal/ApolloHttpLinkOptions.scala
@@ -1,0 +1,10 @@
+package anduin.scalajs.apollo.link.internal
+
+import anduin.scalajs.fetch.internal.FetchFn
+
+import scala.scalajs.js
+
+final class ApolloHttpLinkOptions(
+  val uri: String,
+  val fetch: FetchFn
+) extends js.Object
