+++ modules/apollo-link/src/main/scala/anduin/scalajs/apollo/link/ApolloLink.scala
@@ -1,0 +1,6 @@
+package anduin.scalajs.apollo.link
+
+trait ApolloLink {
+  type Raw <: internal.ApolloLink
+  val raw: Raw
+}
