+++ modules/apollo-link/src/main/scala/anduin/scalajs/apollo/link/internal/ApolloLink.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.apollo.link.internal
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+@JSImport("apollo-link", "ApolloLink", "ApolloLink.ApolloLink")
+@js.native
+class ApolloLink extends js.Object
