+++ modules/node-fetch/src/main/scala/anduin/scalajs/nodefetch/NodeFetch.scala
@@ -1,0 +1,8 @@
+package anduin.scalajs.nodefetch
+
+import anduin.scalajs.fetch.Fetch
+
+object NodeFetch extends Fetch {
+  type Raw = internal.NodeFetch.type
+  val raw: Raw = internal.NodeFetch
+}
